NP-2485 Move translation helpers to own file
