+++ src/pages/registration/contributors_tab/AddContributorModal.tsx
@@ -7,10 +7,7 @@ import Modal from '../../../components/Modal';
 import { RootStore } from '../../../redux/reducers/rootReducer';
 import { Authority } from '../../../types/authority.types';
 import { ContributorRole } from '../../../types/contributor.types';
-import {
-  getAddContributorText,
-  getCreateContributorText,
-} from '../../../utils/validation/registration/contributorTranslations';
+import { getAddContributorText, getCreateContributorText } from '../../../utils/translation-helpers';
 import { AddContributorModalContent } from './components/AddContributorModalContent';
 import { CreateContributorModalContent } from './components/CreateContributorModalContent';
