+++ src/utils/translation-helpers.ts
@@ -1,5 +1,26 @@
-import i18n from '../../../translations/i18n';
-import { ContributorRole } from '../../../types/contributor.types';
+import i18n from '../translations/i18n';
+import { ContributorRole } from '../types/contributor.types';
+import { LanguageCodes } from '../types/language.types';
+import { LanguageString } from '../types/publication_types/commonRegistration.types';
+
+// Map from three letter language to two ("nob" -> "no")
+export const getPreferredLanguageCode = () => {
+  const currentLanguage = i18n.language;
+  if (currentLanguage === LanguageCodes.NORWEGIAN_BOKMAL || currentLanguage === LanguageCodes.NORWEGIAN_NYNORSK) {
+    return 'nb';
+  } else {
+    return 'en';
+  }
+};
+
+// Get label based on selected language
+export const getLanguageString = (labels: LanguageString) => {
+  const preferredLanguageCode = getPreferredLanguageCode();
+  if (Object.keys(labels).includes(preferredLanguageCode)) {
+    return labels[preferredLanguageCode];
+  }
+  return Object.values(labels)[0];
+};
 
 export const getAddContributorText = (contributorRole: string) => {
   switch (contributorRole) {
